backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-socketio>=5.10.0
langchain>=0.1.0
langchain-openai
azure-monitor-opentelemetry-exporter
openinference-instrumentation-langchain 
opentelemetry-sdk 
opentelemetry-exporter-otlp
langgraph
requests>=2.25.0
pytest>=7.0.0
pytest-mock>=3.10.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.2
python-multipart
aiofiles